Refactor from npm workspace to directus bundle extension, switch to pnpm
