--- packages/display/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "directus-extension-display",
-  "version": "1.0.0",
-  "keywords": [
-    "directus",
-    "directus-extension",
-    "directus-custom-display",
-    "directus-field-action-display"
-  ],
-  "directus:extension": {
-    "type": "display",
-    "path": "dist/display-index.js",
-    "source": "src/index.ts",
-    "host": "^9.15.1"
-  },
-  "scripts": {
-    "build": "directus-extension build",
-    "dev": "directus-extension build -w --no-minify",
-    "test": "echo \\\"No display test specified\\\" && exit 0"
-  },
-  "devDependencies": {}
-}
