--- packages/interface/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "directus-extension-directus-extension-text-action-interface",
-  "version": "1.0.0",
-  "keywords": [
-    "directus",
-    "directus-extension",
-    "directus-custom-interface",
-    "directus-field-action-interface"
-  ],
-  "directus:extension": {
-    "type": "interface",
-    "path": "dist/interface-index.js",
-    "source": "src/index.ts",
-    "host": "^9.15.1"
-  },
-  "scripts": {
-    "build": "directus-extension build",
-    "dev": "directus-extension build -w --no-minify",
-    "test": "echo \\\"No interface test specified\\\" && exit 0"
-  },
-  "devDependencies": {}
-}
